Move CI workflow consistency test out of recipe drift into dedicated test suite (#212)
